+++ src/components/Elements/Image/Image.js
@@ -6,11 +6,12 @@ import { BaseProps } from '../../../shared';
 import './image.scss';
 
 /**
- * A lightweight image wrapper to show images in style and responsive
+ * A lightweight image wrapper to show responsive images in style.
  */
-const Image = ({ src, alt, thumb, id, className, style, lazy, ...props }) => {
+const Image = ({ src, alt, thumb, rounded, id, className, style, ...props }) => {
   const classList = classNames('bi bi-image', {
     'bi-img-thumb': thumb,
+    'bi-img-rounded': rounded,
   }, className);
 
   return (<img src={src} alt={alt} id={id} style={style} className={classList} {...props} />);
@@ -30,10 +31,15 @@ Image.propTypes = {
    * Applies the thumbnail style to the image
    */
   thumb: PropTypes.bool,
+  /**
+   * Applies a fully rounded style to the image
+   */
+  rounded: PropTypes.bool,
 };
 
 Image.defaultProps = {
   thumb: false,
+  rounded: false,
 };
 
-export default Image;
+export default React.memo(Image);