+++ src/shared/types/IconProp.js
@@ -1,7 +1,8 @@
 import PropTypes from 'prop-types';
 import Icon from '../../components/Elements/Icon';
+
 /**
- * define type of icon
+ * defines the shared type IconProp
  */
 export default PropTypes.oneOfType([
   PropTypes.string,