+++ CHANGELOG.md
@@ -166,3 +166,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Image component
 - Image component tests
+
+## [0.14.0]  - 2019-07-03
+
+### Added
+
+- `rounded` prop to the Image component
+- Avatar component
+- Avatar component tests
+